fixed a number of bugs with the colour changes
