+++ src/jupyter_contrib_nbextensions/nbextensions/accessibility_toolbar/font_style/font_style.js
@@ -1,10 +1,10 @@
 define([
   "base/js/namespace",
   "jquery",
-  "./font_style/font_control",
-  "./font_style/font_spacing",
-  "./Color_style/Color_control",
-  "./font_style/predefined_styles"
+  "./font_control",
+  "./font_spacing",
+  "./color_control",
+  "./predefined_styles"
 ], function(
   Jupyter,
   $,
@@ -222,6 +222,10 @@ define([
     font_style.prototype.set_default_styles = function() {
       fc_obj.set_default_values();
       fsp_obj.set_default_values();
+      cc_obj.set_colors(
+        cc_obj.background_color_reset(),
+        cc_obj.font_color_reset()
+      );
     };
 
     font_style.prototype.set_saved_style = function() {
@@ -229,6 +233,8 @@ define([
       var saved_letter_space = localStorage.getItem("letter_spacing");
       var saved_font_size = localStorage.getItem("font_size");
       var saved_font_name = localStorage.getItem("font_name");
+      var saved_background_color = localStorage.getItem("background_color");
+      var saved_font_color = localStorage.getItem("font_color");
 
       fc_obj.load_font_change(
         JSON.parse(saved_font_name),
@@ -238,6 +244,11 @@ define([
       fsp_obj.set_line_height(JSON.parse(saved_line_height));
 
       fsp_obj.set_letter_spacing(JSON.parse(saved_letter_space));
+
+      cc_obj.set_colors(
+        JSON.parse(saved_background_color),
+        JSON.parse(saved_font_color)
+      );
     };
   };
